Multi-Cookie Logic (fix)
FIX: original attempt did not account for instances where a single cookie was provided. I have fixed this oversight. However, code is a little redundant. It may be worthwhile to functionalize the cookie split on the "=" and pass the results of the check for the semi-colon character into that function. 

At line 71:

Added logic which check for multiple cookies supplied in the cookie flag. Separation is performed on the semi-colon character. I apologize if I did this wrong.

Original code:

if cookie is not None and len(cookie) > 0:
    if "=" in cookie:
        name = cookie.split("=", 1)[0]
        val = cookie.split("=", 1)[1]
        c = requests.cookies.create_cookie(name=name, value=val)

        _requester.cookies.set_cookie(c)
    else:
        output.error(
            f"Invalid cookie specified ({cookie}) - cookie must be in NAME=VALUE format. Ignored."
        )